change MainActivity.kt
