+++ README.md
@@ -18,3 +18,7 @@
 # 2.1. Обработка событий в Android
 
 - Задача Like, Share
+
+# 2.2. Архитектура: MVVM
+
+- Задача MVVM
